fix circular dependency when importing from component barrel
Importing from the component barrel from within a component that is also in the barrel causes the error.
With optimizePackageImports this wasnt an issue in the app, but storybook didnt work so rather than workaround and intentionally make circular dependencys i decided to add a eslint rule to prevent myself outright.